pyproject.toml
[project]
name = "StaX"
version = "0.1.0"
description = "Open-source advanced solution for mass production stock footage management"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "ffpyplayer>=4.5.3",
    "flask>=3.1.3",
    "imagehash>=4.3.2",
    "numpy>=2.0.2",
    "opencv-python>=4.13.0.92",
    "pillow>=11.3.0",
    "pyinstaller>=6.19.0",
    "pyside2>=5.15.2.1",
    "pytest>=8.4.2",
    "pytest-mock>=3.15.1",
]
